+++ .gitignore
@@ -235,4 +235,4 @@ dmypy.json
 
 
 # Project-level
-/config.py
+**config**
