+++ telegram_bot/__init__.py
@@ -1,0 +1,7 @@
+from ._bot import bot
+# Create `bot` handlers
+from . import message_handlers as _
+from . import callback_handlers as _
+
+
+__all__ = ["bot"]
