+++ telegram_bot/_bot.py
@@ -1,0 +1,6 @@
+import telebot
+
+from .config import bot_token
+
+
+bot = telebot.TeleBot(bot_token)
